+++ R/doc_cran_monthly_package_number.R
@@ -1,11 +1,13 @@
 #' cran_monthly_package_number
 #'
-#' Description.
+#' The evolution of the number of packages on CRAN since its beginning.
+#' Last update: 2022-06-29.
 #'
-#' @format A data frame with 294 rows and 2 variables:
+#' @format A data frame with 297 rows and 2 variables:
 #' \describe{
-#'   \item{ date }{  POSIXct,POSIXt }
-#'   \item{ number_packages }{  numeric }
+#'   \item{ date }{  [Date] Month of release.}
+#'   \item{ number_packages }{  [numeric] Number of packages available on CRAN
+#'   at that given date.}
 #' }
-#' @source Source
+#' @source https://cran.rstudio.com/src/contrib/ and https://mran.microsoft.com/
 "cran_monthly_package_number"